fix:long으eat:follow 작업중, fix:id long으로 변경
Former-commit-id: f53e733373ff0adf5fdb635595c3ddbc381f96cf

+++ Backend/src/main/java/com/ontheblock/www/genre/repository/GenreRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import com.ontheblock.www.genre.domain.Genre;
 
 @Repository
-public interface GenreRepository extends JpaRepository<Genre, Integer> {
+public interface GenreRepository extends JpaRepository<Genre, Long> {
 
 
 }
