+++ Backend/src/main/java/com/ontheblock/www/instrument/repository/InstrumentRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.ontheblock.www.instrument.domain.Instrument;
 
 @Repository
-public interface InstrumentRepository extends JpaRepository<Instrument, Integer> {
+public interface InstrumentRepository extends JpaRepository<Instrument, Long> {
 
 }
